docs/sample-project-1-price-monitor-alerts/requirements.txt
# Core dependencies
uagents>=0.22.3
requests>=2.28.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# For technical analysis
ta>=0.10.0  
